requirements.txt
openai==1.86.0
python-dotenv==1.1.0
openai-agents==0.0.17
sendgrid==6.12.4